+++ hydra-cluster/src/HydraNode.hs
@@ -28,6 +28,7 @@ import Hydra.Options (ChainConfig (..), DirectChainConfig (..), LedgerConfig (..
 import Hydra.Tx.ContestationPeriod (ContestationPeriod)
 import Hydra.Tx.Crypto (HydraKey)
 import Hydra.Tx.DepositDeadline (DepositDeadline)
+import Hydra.Tx.Snapshot
 import Network.HTTP.Conduit (parseUrlThrow)
 import Network.HTTP.Req (GET (..), HttpException, JsonResponse, NoReqBody (..), POST (..), ReqBodyJson (..), defaultHttpConfig, responseBody, runReq, (/:))
 import Network.HTTP.Req qualified as Req
@@ -95,6 +96,14 @@ waitNoMatch delay client match = do
   case result of
     Left _ -> pure () -- Success: waitMatch failed to find a match
     Right _ -> failure "waitNoMatch: A match was found when none was expected"
+
+-- | Wait up to some time and succeed if no API server output matches the given predicate.
+waitForAllNoMatch :: (Eq a, Show a, HasCallStack) => NominalDiffTime -> [HydraClient] -> (Aeson.Value -> Maybe a) -> IO ()
+waitForAllNoMatch delay clients match = do
+  result <- try (void $ waitForAllMatch delay clients match) :: IO (Either SomeException ())
+  case result of
+    Left _ -> pure () -- Success: waitMatch failed to find a match
+    Right _ -> failure "waitForAllNoMatch: A match was found when none was expected"
 
 -- | Wait up to some time for an API server output to match the given predicate.
 waitMatch :: HasCallStack => NominalDiffTime -> HydraClient -> (Aeson.Value -> Maybe a) -> IO a
@@ -215,6 +224,21 @@ getSnapshotUTxO HydraClient{apiHost = Host{hostname, port}} =
       (Req.http hostname /: "snapshot" /: "utxo")
       NoReqBody
       (Proxy :: Proxy (JsonResponse UTxO))
+      (Req.port (fromInteger . toInteger $ port))
+
+-- | Get the latest snapshot from the hydra-node. NOTE: While we usually
+-- avoid parsing responses using the same data types as the system under test,
+-- this parses the response as a 'Snapshot' type as we often need to pick it apart.
+getSnapshot :: HydraClient -> IO (Snapshot Tx)
+getSnapshot HydraClient{apiHost = Host{hostname, port}} =
+  runReq defaultHttpConfig request <&> responseBody
+ where
+  request =
+    Req.req
+      GET
+      (Req.http hostname /: "snapshot")
+      NoReqBody
+      (Proxy :: Proxy (JsonResponse (Snapshot tx)))
       (Req.port (fromInteger . toInteger $ port))
 
 getMetrics :: HasCallStack => HydraClient -> IO ByteString
@@ -406,9 +430,9 @@ withHydraNode' tracer chainConfig workDir hydraNodeId hydraSKey hydraVKeys allNo
   -- NOTE: See comment above about 0.0.0.0 vs 127.0.0.1
   peers =
     [ Host
-        { Network.hostname = "0.0.0.0"
-        , Network.port = fromIntegral $ 5_000 + i
-        }
+      { Network.hostname = "0.0.0.0"
+      , Network.port = fromIntegral $ 5_000 + i
+      }
     | i <- allNodeIds
     , i /= hydraNodeId
     ]