+++ hydra-node/src/Hydra/API/Server.hs
@@ -24,6 +24,8 @@ import Hydra.API.ServerOutput (
   projectHeadStatus,
   projectInitializingHeadId,
   projectPendingDeposits,
+  projectPendingTxs,
+  projectSnapshot,
   projectSnapshotUtxo,
  )
 import Hydra.API.ServerOutputFilter (
@@ -95,18 +97,22 @@ withAPIServer config env party persistence tracer chain pparams serverOutputFilt
     -- NOTE: we do not keep the stored events around in memory
     headStatusP <- mkProjection Idle projectHeadStatus
     snapshotUtxoP <- mkProjection Nothing projectSnapshotUtxo
+    snapshotP <- mkProjection Nothing projectSnapshot
     commitInfoP <- mkProjection CannotCommit projectCommitInfo
     headIdP <- mkProjection Nothing projectInitializingHeadId
     pendingDepositsP <- mkProjection [] projectPendingDeposits
+    pendingTxsP <- mkProjection [] projectPendingTxs
     loadedHistory <-
       runConduitRes $
         source
           -- .| mapC output
           .| iterM (lift . atomically . update headStatusP . output)
           .| iterM (lift . atomically . update snapshotUtxoP . output)
+          .| iterM (lift . atomically . update snapshotP . output)
           .| iterM (lift . atomically . update commitInfoP . output)
           .| iterM (lift . atomically . update headIdP . output)
           .| iterM (lift . atomically . update pendingDepositsP . output)
+          .| iterM (lift . atomically . update pendingTxsP . output)
           -- FIXME: don't load whole history into memory
           .| sinkList
 
@@ -130,7 +136,18 @@ withAPIServer config env party persistence tracer chain pparams serverOutputFilt
             $ websocketsOr
               defaultConnectionOptions
               (wsApp party tracer history callback headStatusP headIdP snapshotUtxoP responseChannel serverOutputFilter)
-              (httpApp tracer chain env pparams (atomically $ getLatest commitInfoP) (atomically $ getLatest snapshotUtxoP) (atomically $ getLatest pendingDepositsP) callback)
+              ( httpApp
+                  tracer
+                  chain
+                  env
+                  pparams
+                  (atomically $ getLatest commitInfoP)
+                  (atomically $ getLatest snapshotUtxoP)
+                  (atomically $ getLatest snapshotP)
+                  (atomically $ getLatest pendingTxsP)
+                  (atomically $ getLatest pendingDepositsP)
+                  callback
+              )
       )
       ( do
           waitForServerRunning
@@ -142,7 +159,9 @@ withAPIServer config env party persistence tracer chain pparams serverOutputFilt
                     update headStatusP output
                     update commitInfoP output
                     update snapshotUtxoP output
+                    update snapshotP output
                     update headIdP output
+                    update pendingTxsP output
                     update pendingDepositsP output
                     writeTChan responseChannel timedOutput
               }